# Downgrade torchmetrics to 0.6.0