--- a/mart/configs/metric/average_precision.yaml
+++ b/mart/configs/metric/average_precision.yaml
@@ -1,11 +1,11 @@
 # @package model
 
 training_metrics:
-  _target_: torchmetrics.MAP
+  _target_: torchmetrics.detection.MAP
   compute_on_step: false
 
 validation_metrics:
-  _target_: torchmetrics.MAP
+  _target_: torchmetrics.detection.MAP
   compute_on_step: false
 
 test_metrics:
@@ -13,7 +13,7 @@ test_metrics:
   _convert_: partial
   metrics:
     map:
-      _target_: torchmetrics.MAP
+      _target_: torchmetrics.detection.MAP
       compute_on_step: false
     json:
       _target_: mart.utils.export.CocoPredictionJSON