# What does this PR do?

torchmetrics > 0.6.0 is slow and this still has not been fixed in newer versions: https://github.com/Lightning-AI/metrics/issues/1024.

This should improve the performance of object detection training from 3 hours to 1 hour.

## Type of change

Please check all relevant options.

- [x] Improvement (non-breaking)
- [ ] Bug fix (non-breaking)
- [ ] New feature (non-breaking)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update

## Testing

I trained an object detector for 50 iterations:
```
$ time CUDA_VISIBLE_DEVICES=1 python -m mart experiment=ArmoryCarlaOverObjDet_TorchvisionFasterRCNN trainer=gpu trainer.precision=16 fit=True trainer.max_steps=50 ++paths.data_dir=/raid/datasets
```

With be5dbcb, it takes:
```
244.83s user 23.34s system 129% cpu 3:26.39 total
```

With 6542ab0, it takes:
```
77.40s user 21.39s system 222% cpu 44.404 total
```

## Before submitting

- [x] The title is **self-explanatory** and the description **concisely** explains the PR
- [x] My **PR does only one thing**, instead of bundling different changes together
- [x] I list all the **breaking changes** introduced by this pull request
- [x] I have commented my code
- [ ] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [x] I have run pre-commit hooks with `pre-commit run -a` command without errors

## Did you have fun?

Make sure you had fun coding 🙃