@dxoigmn 2022-12-09
Note I see the following error:
```
Error executing job with overrides: ['experiment=ArmoryCarlaOverObjDet_TorchvisionFasterRCNN', 'trainer=gpu', 'trainer.precision=16', 'fit=True', 'trainer.max_steps=50', '++paths.data_dir=/raid/datasets']
Traceback (most recent call last):
  File "/home/ccorneli/Projects/MART/mart/__main__.py", line 52, in main
    metric_value = get_metric_value(
  File "/home/ccorneli/Projects/MART/mart/utils/utils.py", line 199, in get_metric_value
    raise Exception(
Exception: Metric value not found! <metric_name=test_metrics/map>
Make sure metric name logged in LightningModule is correct!
Make sure `optimized_metric` name in `hparams_search` config is correct!
```